requirements.txt
Flask
rpi_ws281x
adafruit-circuitpython-neopixel
Pillow
numpy